Media Player Interface
Created the media player interface. There's still stuff to do like rating the song and changing the playlist by the tab on the right, but it's in work condition. There might be some problem with how it looks. On my end it looks fine but it might look different on your end. I added a screen shot of what is should look like.